Fix up front end
